Fixed incorrect order of operations and incorrect sub-command "`systemctl reload`" -> `systemctl daemon-reload`.

+++ audit-n-fix/scripts/kernel/params.sh
@@ -10,14 +10,15 @@ before="$(sysctl -a)"
 	install -m 640 -o 0 -g 0 cnf/sysctl/kernel.conf /etc/sysctl.d/99-security.conf
 #
 # Load Anti-IPv6 sysctl profile
-[[ -f /etc/sysctl.d/99-disable-ipv6.conf ]] ||
+if ! [[ -f /etc/sysctl.d/99-disable-ipv6.conf ]]; then
 	confirm 'Disable IPv6 @ kernel level' &&
-	install -m 640 -o 0 -g 0 cnf/sysctl/kernel-no-ipv6.conf /etc/sysctl.d/99-disable-ipv6.conf
+		install -m 640 -o 0 -g 0 cnf/sysctl/kernel-no-ipv6.conf /etc/sysctl.d/99-disable-ipv6.conf;
+fi
 #
 # Install service to disable mutability of kernel modules 10 seconds after boot.
 confirm 'Disable the loading or unloading of kernel modules 10 seconds after boot' &&
 	install -m 600 -o 0 -g 0 cnf/sysctl/immutable-modules.service /etc/systemd/system/immutable-modules.service &&
-	systemctl reload && systemctl enable --now immutable-modules.service
+	systemctl daemon-reload && systemctl enable --now immutable-modules.service
 #
 # Apply changes
 sysctl --system >/dev/null
